Forward and swerve input implemented
The forward motion is implemented by making the player a child of empty parent and that parent will follow a spline created by gathering the points from the level chunks. The left right movement is implemented on player(not on parent) using the lean gesture.
Parent will control forward motion and turning according to spline.
Child Player will control the rest of maneuvering.